% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inflam_sig}
\alias{inflam_sig}
\title{Inflam Sig}
\format{
An object of class \code{list} of length 1.
}
\usage{
inflam_sig
}
\description{
Inflammation signature from Winslow et. al 2015 Prognostic stromal gene signatures in breast cancer.
}
\keyword{datasets}
