% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metabric_data}
\alias{metabric_data}
\title{METABRIC BRCA Data}
\format{
An object of class \code{ExpressionSet} with 24368 rows and 1980 columns.
}
\usage{
metabric_data
}
\description{
ExpressionSet Object containing 24368 features (HGNC) and 1980 samples. Counts are log transformed
}
\keyword{datasets}
