% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prolif_sig}
\alias{prolif_sig}
\title{Prolif Sig}
\format{
An object of class \code{list} of length 1.
}
\usage{
prolif_sig
}
\description{
Proliferation signature from Venet et. al 2011 Most random gene expression signatures are
significantly associated with breast cancer outcome.
}
\keyword{datasets}
