% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scanb_data}
\alias{scanb_data}
\title{SCANB Data}
\format{
An object of class \code{SummarizedExperiment} with 18129 rows and 9142 columns.
}
\usage{
scanb_data
}
\description{
SummarizedExperiment Object containing 18129 features (HGNC) and 9142 samples. Counts are filtered to remove
low-expressed genes, DESeq2normalized to normalize for library size, and log transformed (log2(x+1)).
}
\keyword{datasets}
