% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_data}
\alias{tcga_data}
\title{TCGA BRCA Data}
\format{
An object of class \code{ExpressionSet} with 37335 rows and 1222 columns.
}
\usage{
tcga_data
}
\description{
ExpressionSet Object containing 37335 features and 1222 samples. Counts are DESeq2normalized.
}
\keyword{datasets}
