% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_data}
\alias{tcga_data}
\title{TCGA BRCA Data}
\format{
An object of class \code{SummarizedExperiment} with 37335 rows and 1222 columns.
}
\usage{
tcga_data
}
\description{
SummarizedExperiment Object containing 37335 features (HGNC) and 1222 samples. Counts are filtered to remove
low-expressed genes, DESeq2normalized to normalize for library size, and log transformed (log2(x+1)).
}
\keyword{datasets}
